backend-flask/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RateLimiter==0.1.0
snowflake-connector-python==3.4.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2





